Changes id column type Integer to BigInteger

+++ backend/app/models/url_model.py
@@ -1,11 +1,11 @@
-from datetime import datetime, timezone, timedelta
-from sqlalchemy import Column, Integer, String, DateTime
+from datetime import datetime, timezone
+from sqlalchemy import Column, Integer, String, DateTime, BigInteger
 from app.db.base import Base
 
 class URL(Base):
     __tablename__ = "urls"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     code = Column(String, unique=True, index=True)
     original_url = Column(String, nullable=False)
     created_at = Column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
